+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 - Updated `pixels` to 0.16.
 - Fixed native startup on Bevy 0.18 by using `RawHandleWrapper` directly instead of the removed
   `WinitWindows` non-send resource.
-- Promote usage of `Single` system params
+- Promote usage of `Single` system params.
+- Integrate Bevy's message API.
 - Add test coverage.
 
 ## [0.15.0] - 2025-03-28